refactored file to lowercase
