+++ applications/models/constants.py
